add actualizar Kilometraje de vehiculos con peticion traccar

+++ src/main/java/com/gifa_api/dto/vehiculo/VehiculoResponseDTO.java
@@ -18,7 +18,7 @@ public class VehiculoResponseDTO {
     private Integer id;
     private String patente;
     private Integer antiguedad;
-    private Integer kilometraje;
+    private double kilometraje;
     private String modelo;
     private EstadoVehiculo estadoVehiculo;
     private EstadoDeHabilitacion estadoDeHabilitacion;
